+++ acme/services/CSE.py
@@ -135,7 +135,7 @@ def startup(args:argparse.Namespace, **kwargs: Dict[str, Any]) -> bool:
 	L.log('Starting CSE')
 	L.log(f'CSE-Type: {cseType.name}')
 	L.log(Configuration.print())
-	L.enableQueue = False		# No queuing of log messages during startup
+	L.queueOff()				# No queuing of log messages during startup
 	
 	# set the logger for the backgroundWorkers. Add an offset to compensate for
 	# this and other redirect functions to determine the correct file / linenumber
@@ -181,7 +181,7 @@ def startup(args:argparse.Namespace, **kwargs: Dict[str, Any]) -> bool:
 		return False 					
 
 	# Enable log queuing
-	L.enableQueue = True	
+	L.queueOn()	
 
 	# Send an event that the CSE startup finished
 	cseStatus = CSEStatus.RUNNING
@@ -224,7 +224,7 @@ def _shutdown() -> None:
 		return
 		
 	cseStatus = CSEStatus.STOPPING
-	L.enableQueue = False
+	L.queueOff()
 	L.isInfo and L.log('CSE shutting down')
 	if event:	# send shutdown event
 		event.cseShutdown() 	# type: ignore
@@ -267,7 +267,7 @@ def resetCSE() -> None:
 		L.enableScreenLogging = Configuration.get('logging.enableScreenLogging')	# Set screen logging to the originally configured values
 
 		L.setLogLevel(Configuration.get('logging.level'))
-		L.enableQueue = False	# Disable log queuing for restart
+		L.queueOff()	# Disable log queuing for restart
 		
 		httpServer.pause()
 		mqttClient.pause()
@@ -285,7 +285,7 @@ def resetCSE() -> None:
 		httpServer.unpause()
 
 		# Enable log queuing again
-		L.enableQueue = True
+		L.queueOn()
 
 		# Send restart event
 		event.cseRestarted()	# type: ignore [attr-defined]